NJTA-10-added-tests-with-coverage

+++ src/main/java/net/njit/ms/cs/repository/BuildingRepository.java
@@ -3,5 +3,5 @@ package net.njit.ms.cs.repository;
 import net.njit.ms.cs.model.entity.Building;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BuildingRepository extends JpaRepository<Building, Long> {
+public interface BuildingRepository extends JpaRepository<Building, Integer> {
 }
